+++ c/src/CMakeLists.txt
@@ -8,8 +8,9 @@ set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${PROJECT_BINARY_DIR}/lib)
 link_directories( ${CMAKE_CURRENT_BINARY_DIR} )
 include_directories (${CMAKE_CURRENT_BINARY_DIR} ${PROJECT_BINARY_DIR}/src )
 
-add_library (variantkey binsearch.c binsearch.h esid.c esid.h genoref.c genoref.h hex.c hex.h nrvk.c nrvk.h regionkey.c regionkey.h rsidvar.c rsidvar.h variantkey.c variantkey.h)
+add_library (variantkey binsearch.h esid.h genoref.h hex.h nrvk.h regionkey.h rsidvar.h variantkey.h)
 target_include_directories (variantkey PUBLIC ${CMAKE_CURRENT_SOURCE_DIR})
+set_target_properties(variantkey PROPERTIES LINKER_LANGUAGE "C")
 
 # Required to link the math library
 target_link_libraries(variantkey)
