+++ c/test/test_esid.c
@@ -44,7 +44,7 @@
 #include <string.h>
 #include <strings.h>
 #include <time.h>
-#include "../src/esid.c"
+#include "../src/esid.h"
 
 // returns current time in nanoseconds
 uint64_t get_time()
