+++ c/test/test_hex.c
@@ -44,7 +44,7 @@
 #include <string.h>
 #include <strings.h>
 #include <time.h>
-#include "../src/hex.c"
+#include "../src/hex.h"
 
 // returns current time in nanoseconds
 uint64_t get_time()
