+++ conda/r/meta.yaml
@@ -1,6 +1,6 @@
 package:
   name: r-variantkey
-  version: 3.0.2
+  version: 4.0.0
 
 source:
   path: ../..
